decouple title into a component
