+++ src/components/v2/Layout/Header/Title/isOnMarketDetailsPage.ts
