+++ src/components/v2/Layout/Header/styles.ts
@@ -25,24 +25,6 @@ export const useStyles = () => {
         padding: ${theme.spacing(6, 4, 0)} !important;
       }
     `,
-    backButton: css`
-      display: flex;
-      align-items: center;
-    `,
-    backButtonChevronIcon: css`
-      width: ${theme.spacing(6)};
-      height: ${theme.spacing(6)};
-      color: ${theme.palette.text.primary};
-      margin-right: ${theme.spacing(2)};
-    `,
-    backButtonTokenIcon: css`
-      width: ${theme.spacing(8)};
-      height: ${theme.spacing(8)};
-      margin-right: ${theme.spacing(2)};
-    `,
-    backButtonTokenSymbol: css`
-      color: ${theme.palette.text.primary};
-    `,
     claimXvsButton: css`
       margin-right: ${theme.spacing(6)};
     `,
@@ -50,6 +32,10 @@ export const useStyles = () => {
       display: flex;
       align-items: center;
       margin-left: auto;
+
+      ${theme.breakpoints.down('md')} {
+        display: none;
+      }
     `,
   };
 };
